# cmd/allternit-cloud-api/openapi/platform-v1.yaml
openapi: 3.1.0
info:
  title: Allternit Platform API
  version: "1.0.0"
  summary: Agents, voice, messaging, numbers, channels and the twin layer, for developers.
  description: |
    The Allternit Platform API. Base URL `https://api.allternit.com`.

    Authenticate with a project API key: `Authorization: Bearer alt_live_…` (live
    project) or `alt_test_…` (sandbox project). Every POST accepts an
    `Idempotency-Key` header (stored per key for 24 hours). Responses carry
    `x-ratelimit-limit`, `x-ratelimit-remaining` and `x-ratelimit-reset` headers.

    Every project, sandbox included, needs a card on file (console → Billing)
    before it can do billable work. Until then billable requests answer
    `402 payment_method_required`, with the console billing page in `error.url`.

    This file is the source of truth for the `/v1` routes: a test fails if the
    router and this file disagree.
servers:
  - url: https://api.allternit.com
security:
  - bearerAuth: []
paths:
...
  /v1/computers:
    get:
      operationId: listComputers
      summary: List computers
      description: Hosted-driver computers in this project (an account-bound key sees its account's only). Requires the `computers` scope and the project's hosted driver flag (404 `hosted_driver_disabled` until Allternit turns it on).
      tags: [Computers]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/After"
      responses:
        "200":
          description: A page of computers.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComputerList" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createComputer
      summary: Create a computer
      description: Provisions a cloud computer your agents drive with the computer and browser toolsets. 402 `payment_method_required` without a card on file, 402 `spend_cap_reached` at the spend cap; 429 `concurrency_limit` when this key already has `per_key_concurrency` live computers.
      tags: [Computers]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string, maxLength: 120 }
                account_id: { type: string }
                metadata: { type: object, additionalProperties: true }
      responses:
        "201":
          description: The computer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Computer" }
        default: { $ref: "#/components/responses/Error" }
  /v1/computers/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      operationId: getComputer
      summary: Get a computer
      tags: [Computers]
      responses:
        "200":
          description: The computer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Computer" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteComputer
      summary: Delete a computer
      description: Deletes the computer and its disk. Running minutes up to now are billed.
      tags: [Computers]
      responses:
        "200":
          description: The deleted computer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Computer" }
        default: { $ref: "#/components/responses/Error" }
...
  /v1/computers/{id}/toolset:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      operationId: callComputerToolset
      summary: Run one toolset call
      description: |
        One `allternit.computer.v1` / `allternit.browser.v1` call (member names and inputs match Anthropic's `computer_toolset_20260801` / `browser_toolset_20260801`). Runs through the computer's executor: contract validation, control lease, policy, approval, audit, then the action. Action failures are 200 with `is_error: true`.
        409 `approval_required` carries `approval`: approve it (`POST /v1/computers/{id}/approvals/{approval_id}`, or the project owner in the console), then resend the same call with `approval_grant`.
      tags: [Computers]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ToolsetCall" }
      responses:
        "200":
          description: The result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolsetResult" }
        "409":
          description: Approval required (or another conflict; then the body is the Error object).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolsetApprovalRequired" }
        default: { $ref: "#/components/responses/Error" }
...
  /v1/computers/{id}/events:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      operationId: listComputerEvents
      summary: Poll the computer's action events
      description: "`computer.action` events, oldest first. Pass the last page's `next_cursor` as `after` to poll."
      tags: [Computers]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/After"
      responses:
        "200":
          description: A page of events.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComputerEventList" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A project API key, `alt_live_…` or `alt_test_…`.
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size, 1 to 100 (default 20).
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    After:
      name: after
      in: query
      description: Opaque cursor from `next_cursor` of the previous page.
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Up to 255 characters. Repeating a request with the same key replays the stored response for 24 hours; the same key with a different request is a 409.
      schema: { type: string, maxLength: 255 }
  responses:
    Error:
      description: Error. The HTTP status says which kind; the body is the Error object.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, code, message, param]
          properties:
            type:
              type: string
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found_error
                - rate_limit_error
                - conflict_error
                - billing_error
                - api_error
            code: { type: string }
            message: { type: string }
            param: { type: [string, "null"] }
            url: { type: string, description: "Where to fix it. Set on 402 payment_method_required: the console billing page for the project." }
...
    Computer:
      type: object
      required: [id, object, name, status, key_id, created_at, metadata]
      properties:
        id: { type: string }
        object: { type: string, const: computer }
        name: { type: string }
        status: { type: string, enum: [provisioning, starting, running, stopping, stopped, error, deleted] }
        account_id: { type: [string, "null"] }
        key_id: { type: string, description: The key that created it (`console` for the console). }
        created_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        metadata: { type: object, additionalProperties: true }
    ComputerList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Computer" } }
        has_more: { type: boolean }
        next_cursor: { type: [string, "null"] }
    ComputerEvent:
      type: object
      properties:
        id: { type: string }
        type: { type: string, const: computer.action }
        ts: { type: string, format: date-time }
        data:
          type: object
          properties:
            toolset: { type: string }
            member: { type: string }
            x: { type: [integer, "null"] }
            y: { type: [integer, "null"] }
            screen_w: { type: [integer, "null"] }
            screen_h: { type: [integer, "null"] }
            run_id: { type: [string, "null"] }
            ok: { type: boolean }
    ComputerEventList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/ComputerEvent" } }
        has_more: { type: boolean }
        next_cursor: { type: [string, "null"] }
...
    ToolsetCall:
      type: object
      required: [toolset, member]
      properties:
        toolset: { type: string, enum: [computer, browser] }
        member: { type: string, description: "A contract member, e.g. screenshot, left_click, type, navigate." }
        input: { type: object, additionalProperties: true }
        run_id: { type: string }
        turn_id: { type: string }
        call_index: { type: integer, minimum: 0 }
        model_frame:
          type: object
          properties:
            width: { type: integer }
            height: { type: integer }
        coordinate_space: { type: string, enum: [pixels, normalized_1000], default: pixels }
        approval_grant: { type: string }
        browser_session_id: { type: string }
        enable:
          type: array
          items: { type: string }
          description: Off-by-default members (file_upload, read_console, read_network, javascript_exec) to allow for this call.
    ToolsetResult:
      type: object
      required: [is_error, content]
      properties:
        is_error: { type: boolean }
        content:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [text, image] }
              text: { type: string }
              media_type: { type: string }
              data: { type: string, description: Base64 image. }
        browser_state: { type: object, additionalProperties: true }
        screen:
          type: object
          properties:
            width: { type: integer }
            height: { type: integer }
            scale: { type: number }
            frame_width: { type: integer }
            frame_height: { type: integer }
        error: { type: string, description: "Machine-readable code on failures: action_failed, unimplemented, not_executed, driver_failed." }
    ToolsetApprovalRequired:
      type: object
      properties:
        error: { type: object, additionalProperties: true, description: "The usual error object, code approval_required." }
        approval:
          type: object
          properties:
            id: { type: string }
            action_hash: { type: string }
            member: { type: string }
            toolset: { type: string }
            risk: { type: string }
            confirmation_class: { type: string }
            approve_url: { type: string }
            approver: { type: string, enum: [owner, api_key] }
        result: { $ref: "#/components/schemas/ToolsetResult" }